Adding new test -- Component is coming close to be finished
NOTE: The new test contains 4 expect clauses. This can be a little bit hard to debug once you start building code to try this test to pass, as while it's failing, it may be hard to know on which expect it is actualling failing at.

Alternatively, the test can be broken into 2 tests:

1) Check that the active tile has a border (2 expects)
2) Check that the inactive tiles have no border (2 expects)

The way you write your tests depends entirely on what you feel makes the most sense.