Adding final test
Note: In this test, we use a spy to simulate a function execution. This method can be also used to simulate asyn calls (for example, if testing a redux store). Using spys makes it so that our tests don't rely on a function execution or async reponse to happen, making sure our tests run fast and smooth

At the same time, we use our test to define our component's API, in this case, declaring what the new prop for the callback function should be called.